requirements.txt
flask
flask-session
flask-wtf
# flask-mysqldb
flask-sqlalchemy
flask-login
flask-admin
flask-migrate
flask-cors
mysql-connector-python
gunicorn
WTForms
mysql-connector
passlib
python-dotenv
pymysql
redis
pytest
pytest-flask
pytest-cov
waitress